requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dialogflow==0.7.2
Flask==1.1.1
future==0.18.2
fuzzywuzzy==0.17.0
google-api-core==1.15.0
google-auth==1.10.0
googleapis-common-protos==1.6.0
grpcio==1.26.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
line-bot-sdk==1.15.0
MarkupSafe==1.1.1
panda3d==1.10.4.1
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-Levenshtein-wheels==0.13.1
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
